DRTII-1122 Tidying
