+++ .drone-v1.yml
@@ -100,9 +100,6 @@ steps:
       SLACK_WEBHOOK:
         from_secret: slack_webhook
     when:
-#      status:
-#        - success
-#        - failure
       event:
         - push
     depends_on:
