+++ src/main/scala/uk/gov/homeoffice/drt/analytics/AnalyticsApp.scala
@@ -12,7 +12,7 @@ import uk.gov.homeoffice.drt.analytics.prediction.modeldefinitions._
 import uk.gov.homeoffice.drt.analytics.prediction.{FlightRouteValuesTrainer, ModelDefinition}
 import uk.gov.homeoffice.drt.analytics.s3.Utils
 import uk.gov.homeoffice.drt.analytics.services.ArrivalsHelper.{noopPreProcess, populateMaxPax}
-import uk.gov.homeoffice.drt.analytics.services.ModelAccuracy
+import uk.gov.homeoffice.drt.analytics.services.{ModelAccuracy, PaxModelStats}
 import uk.gov.homeoffice.drt.arrivals.Arrival
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Terminals.Terminal
@@ -71,11 +71,11 @@ object AnalyticsApp extends App {
 
         case "update-pax-cap-models" =>
           trainModels(PaxCapModelDefinition, portConfig.terminals, populateMaxPax()).flatMap { _ =>
-            ModelAccuracy.analyse(daysOfTrainingData, portCode.iata, portConfig.terminals, PaxCapModelStats, paxCapModelCollector, bucketName)
+            ModelAccuracy.analyse(daysOfTrainingData, portCode.iata, portConfig.terminals, paxCapModelCollector, bucketName)
           }
 
         case "dump-daily-pax-cap" =>
-          ModelAccuracy.analyse(daysOfTrainingData, portCode.iata, portConfig.terminals, PaxCapModelStats, paxCapModelCollector, bucketName)
+          ModelAccuracy.analyse(daysOfTrainingData, portCode.iata, portConfig.terminals, paxCapModelCollector, bucketName)
 
         case unknown =>
           log.error(s"Unknown job name '$unknown'")
